[TASK] Use TemplateDao for sys-template queries
Issue #157
